Remove unnecessary mappers
