+++ common/src/main/java/net/draycia/carbon/common/users/db/mysql/ComponentArgumentFactory.java
@@ -17,24 +17,28 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package net.draycia.carbon.common.users.db.jdbi;
+package net.draycia.carbon.common.users.db.mysql;
 
 import java.sql.Types;
 import java.util.UUID;
+import net.kyori.adventure.key.Key;
+import net.kyori.adventure.text.Component;
+import net.kyori.adventure.text.serializer.gson.GsonComponentSerializer;
 import org.jdbi.v3.core.argument.AbstractArgumentFactory;
 import org.jdbi.v3.core.argument.Argument;
 import org.jdbi.v3.core.config.ConfigRegistry;
 
-// https://github.com/broccolai/tickets/blob/55a23b5fcfdc8e4b4bfa861ea7ec620506bd0dfa/core/src/main/java/broccolai/tickets/core/storage/factory/UUIDArgumentFactory.java
-public final class UUIDArgumentFactory extends AbstractArgumentFactory<UUID> {
+public final class ComponentArgumentFactory extends AbstractArgumentFactory<Component> {
 
-    public UUIDArgumentFactory() {
+    final GsonComponentSerializer serializer = GsonComponentSerializer.gson();
+
+    public ComponentArgumentFactory() {
         super(Types.VARCHAR);
     }
 
     @Override
-    public Argument build(final UUID value, final ConfigRegistry config) {
-        return (position, statement, ctx) -> statement.setString(position, value.toString());
+    public Argument build(final Component value, final ConfigRegistry config) {
+        return (position, statement, ctx) -> statement.setString(position, this.serializer.serialize(value));
     }
 
 }
