+++ common/src/main/java/net/draycia/carbon/common/users/db/mysql/UUIDArgumentFactory.java
@@ -17,23 +17,23 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package net.draycia.carbon.common.users.db.jdbi;
+package net.draycia.carbon.common.users.db.mysql;
 
 import java.sql.Types;
 import java.util.UUID;
-import net.kyori.adventure.key.Key;
 import org.jdbi.v3.core.argument.AbstractArgumentFactory;
 import org.jdbi.v3.core.argument.Argument;
 import org.jdbi.v3.core.config.ConfigRegistry;
 
-public final class KeyArgumentFactory extends AbstractArgumentFactory<Key> {
+// https://github.com/broccolai/tickets/blob/55a23b5fcfdc8e4b4bfa861ea7ec620506bd0dfa/core/src/main/java/broccolai/tickets/core/storage/factory/UUIDArgumentFactory.java
+public final class UUIDArgumentFactory extends AbstractArgumentFactory<UUID> {
 
-    public KeyArgumentFactory() {
+    public UUIDArgumentFactory() {
         super(Types.VARCHAR);
     }
 
     @Override
-    public Argument build(final Key value, final ConfigRegistry config) {
+    public Argument build(final UUID value, final ConfigRegistry config) {
         return (position, statement, ctx) -> statement.setString(position, value.toString());
     }
